refactor: user slices
