+++ frontend/zillow/src/store/index.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import GlobalSlice from "./slices/sidebar/sidebarSlice";
-import userReducer from "../slices/user/userSlice";
+import userReducer from "./slices/user/userSlice";
 
 export default configureStore({
   reducer: {
